♻️ Move DB class into the db.py module
